正在编码清册入库

+++ src/main/java/com/gqj/entity/Tool.java
@@ -26,7 +26,7 @@ public class Tool extends Base {
 
 	private Date toolRejectDate;
 
-	private BigDecimal toolTestDateCircle;
+	private Double toolTestDateCircle;
 
 	private String toolRemark;
 
@@ -47,8 +47,7 @@ public class Tool extends Base {
 	}
 
 	public void setToolCode(String toolCode) {
-		this.toolCode = toolCode == null ? null
-				: toolCode.trim();
+		this.toolCode = toolCode == null ? null : toolCode.trim();
 	}
 
 	public Long getStoreId() {
@@ -80,8 +79,7 @@ public class Tool extends Base {
 	}
 
 	public void setToolBox(String toolBox) {
-		this.toolBox = toolBox == null ? null
-				: toolBox.trim();
+		this.toolBox = toolBox == null ? null : toolBox.trim();
 	}
 
 	public Long getToolStatus() {
@@ -116,12 +114,11 @@ public class Tool extends Base {
 		this.toolRejectDate = toolRejectDate;
 	}
 
-	public BigDecimal getToolTestDateCircle() {
+	public Double getToolTestDateCircle() {
 		return toolTestDateCircle;
 	}
 
-	public void setToolTestDateCircle(
-			BigDecimal toolTestDateCircle) {
+	public void setToolTestDateCircle(Double toolTestDateCircle) {
 		this.toolTestDateCircle = toolTestDateCircle;
 	}
 
@@ -130,8 +127,7 @@ public class Tool extends Base {
 	}
 
 	public void setToolRemark(String toolRemark) {
-		this.toolRemark = toolRemark == null ? null
-				: toolRemark.trim();
+		this.toolRemark = toolRemark == null ? null : toolRemark.trim();
 	}
 
 	public Date getToolNextTestDate() {
